vector-sanity/requirements.txt
pyyaml
regex
rich
